Create idk what I am doing both in this code and on github
I am trying to make a program where a teacher can be prompted to enter in information about students in the console and then later on can request info about the student and the console will give it to them. The site I use to work my programs on is called GDB online debugger just so you know. Maybe someone could help? Idk